ai_agents/requirements.txt
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
twilio==9.2.4
python-dotenv==1.0.1
pydantic==2.9.0
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.24.0
black==24.8.0
